@@ -22,7 +22,7 @@
         }
     ],
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.4",
         "ext-json": "*",
         "monolog/monolog": "^2.8",
         "php-di/php-di": "^7.0.7",
@@ -30,10 +30,10 @@
         "slim/slim": "^4.10"
     },
     "require-dev": {
-        "jangregor/phpstan-prophecy": "^1.0.0",
-        "phpspec/prophecy-phpunit": "^2.0",
+        "jangregor/phpstan-prophecy": "^2.0.0",
+        "phpspec/prophecy-phpunit": "^v2.3.0",
         "phpstan/extension-installer": "^1.2.0",
-        "phpstan/phpstan": "^1.8",
+        "phpstan/phpstan": "^2.0",
         "phpunit/phpunit": "^9.5.26",
         "squizlabs/php_codesniffer": "^3.7"
     },
