@@ -2,8 +2,6 @@
 
 declare(strict_types=1);
 
-use App\Application\Actions\User\ListUsersAction;
-use App\Application\Actions\User\ViewUserAction;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Slim\App;
@@ -20,13 +18,7 @@ return function (App $app) {
         return $response;
     });
 
-    $app->get('/a', function (Request $request, Response $response) {
-        $response->getBody()->write('Hello world!!!!');
-        return $response;
-    });
-
-    $app->group('/users', function (Group $group) {
-        $group->get('', ListUsersAction::class);
-        $group->get('/{id}', ViewUserAction::class);
+    $app->group('/selling_targets', function (Group $group) {
+        $group->get('/{id}', \App\Application\Actions\SellingTargets\GetIdAction::class);
     });
 };
